refactor: add index.js files to components folders for shortcut imports
