+++ packages/design-system/components/badge/__stories__/Badge.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Canvas, Story, ArgsTable } from "@storybook/addon-docs"
 
-import Badge from "../Badge.jsx"
+import { Badge } from ".."
 
 <Meta title="Components/Badge" component={Badge} />
 
@@ -8,7 +8,7 @@ import Badge from "../Badge.jsx"
 
 <Canvas>
   <Story
-    name="Example"
+    name="Overview"
     args={{
       text: "Szczeniak ",
       variant: "gray",
