+++ packages/design-system/components/badge/index.js
@@ -1,0 +1,1 @@
+export { default as Badge } from "./Badge"
