+++ packages/design-system/components/banner/__stories__/Banner.stories.mdx
@@ -1,8 +1,8 @@
 import { Meta, Canvas, Story, ArgsTable } from "@storybook/addon-docs"
 
-import Button from "../../button/Button"
+import { Button } from "../../button"
 
-import Banner from "../Banner"
+import { Banner } from ".."
 
 <Meta title="Components/Banner" component={Banner} />
