+++ packages/design-system/components/banner/index.js
@@ -1,0 +1,1 @@
+export { default as Banner } from "./Banner"
