+++ packages/design-system/components/button/__stories__/Button.stories.mdx
@@ -1,7 +1,7 @@
 import { Meta, Canvas, Story, ArgsTable } from "@storybook/addon-docs"
 
-import Button from "../Button"
-import Icon from "../../icon/Icon"
+import { Button } from ".."
+import { Icon } from "../../icon"
 
 <Meta title="Components/Button" component={Button} />
