+++ packages/design-system/components/icon/__stories__/Icon.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Canvas, Story, ArgsTable } from "@storybook/addon-docs"
 
-import Icon from "../Icon"
+import { Icon } from ".."
 
 <Meta title="Components/Icon" component={Icon} />
