+++ packages/design-system/components/icon/index.js
@@ -1,0 +1,2 @@
+export { default as Icon } from "./Icon"
+export { icons } from "./Icon"
