+++ packages/design-system/components/list/__stories__/List.stories.mdx
@@ -1,7 +1,7 @@
 import { Meta, Canvas, Story, ArgsTable } from "@storybook/addon-docs"
 
-import List from "../List"
-import ListItem from "../../listItem/ListItem"
+import { List } from ".."
+import { ListItem } from "../../listItem"
 
 <Meta title="Components/List" component={List} />
