+++ packages/design-system/components/list/index.js
@@ -1,0 +1,1 @@
+export { default as List } from "./List"
