+++ packages/design-system/components/listItem/__stories__/ListItem.stories.mdx
@@ -1,7 +1,7 @@
 import { Meta, Canvas, Story, ArgsTable } from "@storybook/addon-docs"
 
-import ListItem from "../ListItem.jsx"
-import Icon from "../../icon/Icon"
+import { ListItem } from ".."
+import { Icon } from "../../icon"
 
 <Meta title="Components/ListItem" component={ListItem} />
