+++ packages/design-system/components/listItem/index.js
@@ -1,0 +1,1 @@
+export { default as ListItem } from "./ListItem"
