+++ packages/design-system/components/title/__stories__/Ttitle.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Canvas, Story, ArgsTable } from "@storybook/addon-docs"
 
-import Title from "../Title"
+import { Title } from ".."
 
 <Meta title="Components/Title" component={Title} />
