+++ packages/design-system/components/title/index.js
@@ -1,0 +1,1 @@
+export { default as Title } from "./Title"
