+++ packages/design-system/components/typography/__stories__/Typography.stories.mdx
@@ -1,7 +1,7 @@
 import { Meta, Canvas, Story, ArgsTable } from "@storybook/addon-docs"
 import { useTheme } from "@emotion/react"
 
-import Typography from "../Typography"
+import { Typography } from ".."
 
 <Meta title="Components/Typography" component={Typography} />
