+++ packages/design-system/icons/__stories__/Icons.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Typeset, IconGallery, IconItem } from "@storybook/addon-docs"
 
-import { icons } from "../../components/icon/Icon.jsx"
+import { icons } from "../../components/icon"
 
 <Meta title="Icons" />
