+++ packages/design-system/theme/__stories__/Theme.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, ColorPalette, ColorItem, Typeset } from "@storybook/addon-docs"
 
-import theme from "../theme"
+import { theme } from ".."
 
 <Meta
   title="Theme"
