+++ packages/design-system/theme/index.js
@@ -1,0 +1,1 @@
+export { default as theme } from "./theme"
